# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_jumayevgadam_tsu-toleg_internal_models_faculty.Faculty:
    properties:
      faculty-code:
        type: string
      facultyID:
        type: integer
      facultyName:
        type: string
    required:
    - faculty-code
    - facultyName
    type: object
  github_com_jumayevgadam_tsu-toleg_internal_models_group.GroupDTO:
    properties:
      classCode:
        type: string
      facultyID:
        type: integer
      groupID:
        type: integer
    type: object
...
host: localhost:4000
info:
  contact:
    email: hypergadam@gmail.com
    name: Gadam Jumayev
    url: https://github.com/jumayevgadam
  description: This is the API for the TSU-TOLEG system.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms
  title: TSU-TOLEG API Documentation
  version: "2.0"
paths:
  /faculty/{id}:
    delete:
      description: delete faculty using identified id.
      operationId: delete-faculty
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: successfully deleted faculty
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: delete-faculty
      tags:
      - Faculties
    get:
      consumes:
      - multipart/form-data
      description: retrieve faculty using its id.
      operationId: get-faculty
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jumayevgadam_tsu-toleg_internal_models_faculty.Faculty'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: get-faculty.
      tags:
      - Faculties
    put:
      consumes:
      - multipart/form-data
      description: update faculty fields using identified faculty id
      operationId: update-faculty
      parameters:
      - description: needed faculty id for update
        in: path
        name: id
        required: true
        type: integer
      - in: formData
        name: faculty-code
        type: string
      - in: formData
        name: faculty-name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: successfully updated faculty ops
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Update Faculty
      tags:
      - Faculties
  /faculty/create:
    post:
      consumes:
      - multipart/form-data
      description: Creates a new faculty and returns its id.
      operationId: add-faculty
      parameters:
      - in: formData
        name: faculty-code
        required: true
        type: string
      - in: formData
        name: faculty-name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Add a new faculty.
      tags:
      - Faculties
  /faculty/get-all:
    get:
      description: list faculties, pagination not setted.
      operationId: list-faculties
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_jumayevgadam_tsu-toleg_internal_models_faculty.Faculty'
            type: array
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: List faculties.
      tags:
      - Faculties
  /group/{id}:
    delete:
      consumes:
      - multipart/form-data
      description: Delete group by id
      operationId: delete-group
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Delete group.
      tags:
      - Groups
    get:
      consumes:
      - multipart/form-data
      description: Retrieve a group using by identified id.
      operationId: get-group
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jumayevgadam_tsu-toleg_internal_models_group.GroupDTO'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get one group by its id.
      tags:
      - Groups
    put:
      consumes:
      - multipart/form-data
      description: update a group details using their fields.
      operationId: update-group
      parameters:
      - description: needed group id for update
        in: path
        name: id
        required: true
        type: integer
      - in: formData
        name: faculty-id
        type: integer
      - in: formData
        name: group-code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: updated group
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Update a group.
      tags:
      - Groups
  /group/add:
    post:
      consumes:
      - multipart/form-data
      description: Creates a new group and returns its id.
      operationId: add-group
      parameters:
      - in: formData
        name: faculty-id
        required: true
        type: integer
      - in: formData
        name: group-code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Add a new group.
      tags:
      - Groups
  /group/get-all:
    get:
      consumes:
      - multipart/form-data
      description: Listing groups by pagination.
      operationId: list-group
      parameters:
      - description: page number
        format: page
        in: query
        name: page
        type: integer
      - description: number of elements per page
        format: limit
        in: query
        name: limit
        type: integer
      - description: filter name
        format: orderBy
        in: query
        name: orderBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_jumayevgadam_tsu-toleg_internal_models_group.GroupDTO'
            type: array
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: List groups.
      tags:
      - Groups
...
swagger: "2.0"